serverless-log-insights/api/requirements.txt
azure-functions
azure-data-tables==12.8.0
